Finish the mint function

+++ src/DecentralizedStableCoin.sol
@@ -21,6 +21,7 @@
 // view & pure functions
 
 pragma solidity ^0.8.19;
+
 import {ERC20Burnable, ERC20} from "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
@@ -41,7 +42,7 @@ contract DecentralizedStableCoin is ERC20Burnable, Ownable {
     constructor() ERC20("DecentralizedStableCoin", "DSC") Ownable(msg.sender) {}
 
     function mint(uint256 _amount) external onlyOwner {
-        uint balance = balanceOf(msg.sender);
+        uint256 balance = balanceOf(msg.sender);
         if (_amount <= 0) {
             revert DecentralizedStableCoin__MustBeMoreThanZero();
         }
@@ -51,10 +52,7 @@ contract DecentralizedStableCoin is ERC20Burnable, Ownable {
         super.burn(_amount);
     }
 
-    function mint(
-        address _to,
-        uint256 _amount
-    ) external onlyOwner returns (bool) {
+    function mint(address _to, uint256 _amount) external onlyOwner returns (bool) {
         if (_to == address(0)) {
             revert DecentralizedStableCoin__NotZeroAddress();
         }
